requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==3.2.10
django-autocomplete-light==3.9.1
django-el-pagination==3.3.0
django-mail-queue==3.2.4
django-registration-redux==2.10
django-versatileimagefield==2.2
engineering-notation==0.6.0
gunicorn==20.1.0
idna==3.3
Pillow==9.0.1
psycopg2-binary==2.9.3
pycparser==2.21
python-decouple==3.5
python-magic==0.4.25
pytz==2021.3
razorpay==1.3.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tk-tools==0.14.0
urllib3==1.26.9
xlrd==1.2.0